changed x and y from float to double

+++ src/main/java/engine/data/Position.java
@@ -1,4 +1,4 @@
 package engine.data;
 
-public record Position(float x, float y) {
+public record Position(double x, double y) {
 }
